Make vision into an impl/abstract class system
