+++ src/main/java/com/stuypulse/robot/subsystems/vision/LimelightVision.java
@@ -17,22 +17,11 @@ import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
-public class LimelightVision extends SubsystemBase {
-    private static final LimelightVision instance;
-    private final MegaTagMode megaTagMode = MegaTagMode.MEGATAG1;
-    private final int maxTagCount = 2;
-    private boolean doRejectUpdate = false;
-    private boolean apriltagDetected = false;
-    private Pose2d limelightPose;
-    private final Matrix<N3, N1> visionStdDevs;
+public abstract class LimelightVision extends SubsystemBase {
+    public static final LimelightVision instance;
     
-    private final Field2d field = new Field2d();
-
-    private final DifferentialDrivePoseEstimator poseEstimator;
-    private final Drivetrain drivetrain;
-
     static {
-        instance = new LimelightVision();
+        instance = new LimelightVisionImpl();
     }
 
     public static LimelightVision getInstance() {
@@ -44,103 +33,7 @@ public class LimelightVision extends SubsystemBase {
         MEGATAG2
     }
 
-    private LimelightVision() {
-        visionStdDevs = VecBuilder.fill(.1, .1, .1);
+    public abstract Pose2d getEstimatedPose();
 
-        for (Camera camera : Cameras.LimelightCameras) {
-            Pose3d robotRelativePose = camera.getLocation();
-            LimelightHelpers.setCameraPose_RobotSpace(
-                    camera.getName(),
-                    robotRelativePose.getX(),
-                    -robotRelativePose.getY(),
-                    robotRelativePose.getZ(),
-                    Units.radiansToDegrees(robotRelativePose.getRotation().getX()),
-                    Units.radiansToDegrees(robotRelativePose.getRotation().getY()),
-                    Units.radiansToDegrees(robotRelativePose.getRotation().getZ()));
-        }
-
-        drivetrain = Drivetrain.getInstance();
-        poseEstimator = new DifferentialDrivePoseEstimator(drivetrain.getKinematics(), 
-            drivetrain.getHeading(),
-            drivetrain.getLeftDistance(),
-            drivetrain.getRightDistance(),
-            drivetrain.getPose());
-    }
-     
-    private void updatePoseEstimator() {
-        apriltagDetected = false;
-        doRejectUpdate = false;
-
-        poseEstimator.update(drivetrain.getHeading(), 
-            Units.metersToInches(drivetrain.getLeftDistance()), 
-            Units.metersToInches(drivetrain.getRightDistance()));
-
-        if (megaTagMode == MegaTagMode.MEGATAG1) {
-            LimelightHelpers.PoseEstimate mt1 = LimelightHelpers.getBotPoseEstimate_wpiBlue("limelight");
-
-            if (mt1.tagCount == 1 && mt1.rawFiducials.length == 1) {
-                if(mt1.rawFiducials[0].ambiguity > .7) {
-                    doRejectUpdate = true;
-                }
-                if(mt1.rawFiducials[0].distToCamera > 3) {
-                    doRejectUpdate = true;
-                }
-            }
-
-            if (mt1.tagCount == 0) {
-                doRejectUpdate = true;
-            }
-
-            if(!doRejectUpdate) {
-                apriltagDetected = true;
-
-                poseEstimator.setVisionMeasurementStdDevs(visionStdDevs);
-                poseEstimator.addVisionMeasurement(
-                    mt1.pose,
-                    mt1.timestampSeconds);
-            }
-
-        } else if (megaTagMode == MegaTagMode.MEGATAG2) {
-            LimelightHelpers.SetRobotOrientation("limelight", poseEstimator.getEstimatedPosition().getRotation().getDegrees(), 0,0, 0, 0,0);
-            LimelightHelpers.PoseEstimate mt2 = LimelightHelpers.getBotPoseEstimate_wpiBlue_MegaTag2("limelight");
-            
-            if(Math.abs(Drivetrain.getInstance().getGyroRate()) > 360) {
-                doRejectUpdate = true;
-            }
-
-            if(mt2.tagCount == 0) {
-                doRejectUpdate = true;
-            }
-
-            if(!doRejectUpdate) {
-                apriltagDetected = true;
-
-                poseEstimator.setVisionMeasurementStdDevs(visionStdDevs);
-                poseEstimator.addVisionMeasurement(
-                    mt2.pose,
-                    mt2.timestampSeconds
-                );             
-            }
-        }
-    }
-
-    public Pose2d getEstimatedPose() {
-        updatePoseEstimator();
-        limelightPose = poseEstimator.getEstimatedPosition();
-        return limelightPose;
-    }
-
-    public void resetEstimatedPose(Pose2d newPose) {
-        poseEstimator.resetPose(newPose);
-        getEstimatedPose();
-    }
-
-    @Override
-    public void periodic() {
-        getEstimatedPose();
-        field.setRobotPose(limelightPose);
-
-        SmartDashboard.putData("Field", field);
-        SmartDashboard.putBoolean("Vision/AprilTag Detected?", apriltagDetected);
-     }
+    public abstract void resetEstimatedPose(Pose2d newPose);
 }
